ext4_utils: Enable encryption feature for the disk image
* Add FEATURE_INCOMPAT_ENCRYPT flag
* Enable encryption feature for the disk image
* Change feat_incompat member in fs_info structure

The latest versions of Linux kernel require FEATURE_INCOMPAT_ENCRYPT
flag set in the disk image superblock for enabling file-encryption
support.
Without this flag, Android will not be able to set encryption policy.

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/fs/ext4/ioctl.c?id=refs/tags/v4.11#n838

Test: build and flash

+++ ext4_utils/include/ext4_utils/ext4_sb.h
@@ -41,7 +41,7 @@ struct fs_info {
 	uint32_t journal_blocks;
 	uint16_t feat_ro_compat;
 	uint16_t feat_compat;
-	uint16_t feat_incompat;
+	uint32_t feat_incompat;
 	uint32_t bg_desc_reserve_blocks;
 	const char *label;
 	uint8_t no_journal;
